feature: moved more stuff out of portal project
