+++ Datahub.Portal.Metadata/Components/CardSkeleton.razor
