+++ Datahub.Portal.Metadata/_Imports.razor
@@ -37,24 +37,9 @@
 @using Datahub.Core.Layout
 @using Datahub.Achievements
 
-
-@using Datahub.GeoCore.Service
-
 @using Datahub.Portal.Services.Storage
 @using Datahub.Portal.Services
 @using Datahub.Portal.Components
-@using Datahub.Portal.Components.FileGridComponents
-@using Datahub.Portal.Metadata.Components
-@using Datahub.Portal.Views
-@using Datahub.Portal.Views.Modals
-@using Datahub.Portal.Components.Versions
-@using Datahub.Portal.Components.Forms
-@using Datahub.Portal.Data.Finance;
-@using Datahub.Portal.Data;
-@using Datahub.Portal.Data.LanguageTraining;
-@using Datahub.Portal.Data.WebAnalytics;
-
-@using Datahub.LanguageTraining;
 
 @using Elemental.Components;
