+++ Datahub.PowerBI/Datahub.PowerBI.csproj
@@ -1,14 +1,22 @@
-<Project Sdk="Microsoft.NET.Sdk">
+<Project Sdk="Microsoft.NET.Sdk.Razor">
 
   <PropertyGroup>
     <TargetFramework>net7.0</TargetFramework>
     <ImplicitUsings>enable</ImplicitUsings>
     <Nullable>enable</Nullable>
   </PropertyGroup>
+
+ 
     <ItemGroup>
 	    <PackageReference Include="Microsoft.PowerBI.Api" Version="4.11.0" />
     </ItemGroup>
     <ItemGroup>
       <ProjectReference Include="..\Datahub.Core\Datahub.Core.csproj" />
+      <ProjectReference Include="..\Datahub.Portal.Metadata\Datahub.Portal.Metadata.csproj" />
+    </ItemGroup>
+    <ItemGroup>
+      <Content Update="Pages\Admin\PowerBIAdmin.razor">
+        <ExcludeFromSingleFile>true</ExcludeFromSingleFile>
+      </Content>
     </ItemGroup>
 </Project>
