+++ Datahub.PowerBI/Pages/Admin/PowerBIAdmin.razor
@@ -6,7 +6,6 @@
 @page "/admin/powerbi/{ProjectAcronym}/report/{ReportId}"
 @page "/admin/powerbi/{ProjectAcronym}/dataset/{DatasetId}"
 @page "/admin/powerbi/{ProjectAcronym}/workspace/{WorkspaceId}"
-@using Datahub.Metadata.Model
 
 @inject IPowerBiDataService _powerBiDataService
 @inject PowerBiServiceApi _powerBiServiceApi
