+++ Datahub.PowerBI/Pages/Admin/PowerBiManagementTree.razor
@@ -1,4 +1,4 @@
-﻿@using Datahub.Portal.Data
+﻿@using Datahub.PowerBI.Data;
 
 @*TODO: override icons in AeTreeView with FA icons instead of this hack*@
 <style>
