+++ Datahub.PowerBI/Pages/InternalPublic/InternalPowerBIEmbeddedPage.razor
@@ -3,7 +3,7 @@
 @inject IJSRuntime JSRuntime
 @inject IConfiguration _configuration
 
-@inject IStringLocalizer<DataLabels> DataLocalizer
+@inject IStringLocalizer<InternalPowerBIEmbeddedPage> DataLocalizer
 @inject NavigationManager NavigationManager
 @inject PowerBiServiceApi PowerBiServiceApi
 @using Microsoft.Identity.Web
