+++ Datahub.PowerBI/Pages/PowerBI/InternalPublishedPowerBiReportEmbeddedPage.razor
@@ -1,5 +1,6 @@
 ﻿@page "/internal-published-report/{ObjectId}"
 @page "/internal-published-report/{ObjectId}/{Language}"
+@using Datahub.PowerBI.Data;
 @using Microsoft.Identity.Web
 @using LanguageEnum = Metadata.Model.CatalogObjectLanguage
 @using CatalogItem = Metadata.DTO.CatalogObjectResult
