+++ Datahub.PowerBI/Pages/PowerBI/PowerBIEmbeddedPage.razor
@@ -5,7 +5,7 @@
 @inject IJSRuntime JSRuntime
 @inject IConfiguration _configuration
 
-@inject IStringLocalizer<DataLabels> DataLocalizer
+@inject IStringLocalizer<PowerBIEmbeddedPage> DataLocalizer
 @inject NavigationManager NavigationManager
 @inject PowerBiServiceApi PowerBiServiceApi
 @using Microsoft.Identity.Web
