+++ Datahub.PowerBI/Pages/PowerBI/PowerBIPage.razor
@@ -1,5 +1,5 @@
 ﻿@page "/powerbi"
-@inject IStringLocalizer<DataLabels> DataLocalizer
+@inject IStringLocalizer<PowerBIEmbeddedPage> DataLocalizer
 @inject NavigationManager NavigationManager
 @inject IJSRuntime JsInterop
 @inject PowerBiServiceApi PowerBiServiceApi
