+++ Datahub.PowerBI/Pages/PowerBI/PowerBiReportMetadataEditor.razor
@@ -1,4 +1,6 @@
 ﻿@using Datahub.Metadata.Model
+@using Datahub.Portal.Metadata.Components;
+@using Datahub.PowerBI.Data;
 @inject IPowerBiDataService _powerBiDataService
 @inject TranslationService _translationService
 @inject IMetadataBrokerService _metadataService
