+++ Datahub.PowerBI/Pages/Project/ApprovePowerBiRequest.razor
@@ -1,4 +1,5 @@
 ﻿@page "/project/{ProjectAcronym}/powerbirequest"
+@using Datahub.PowerBI.Data;
 
 @inject IDbContextFactory<DatahubProjectDBContext> _contextFactory
 @inject PowerBiServiceApi _powerBiApi
