+++ Datahub.PowerBI/Pages/Public/PowerBIEmbeddedExternal.razor
@@ -3,7 +3,7 @@
 @inject IDbContextFactory<DatahubProjectDBContext> _dbFactoryProject
 @layout PublicLayout
 @inject NavigationManager NavManager
-@using Datahub.Portal.Layout
+@using Datahub.Core.Layout
 @inject IJSRuntime JSRuntime
 @inject IConfiguration _configuration
 @using Microsoft.AspNetCore.Cryptography.KeyDerivation;
