+++ Datahub.PowerBI/_Imports.razor
@@ -20,6 +20,11 @@
 @using Tewr.Blazor.FileReader
 @using Microsoft.Extensions.Hosting
 
+@using Datahub.PowerBI.Data;
+@using Datahub.Metadata.Model
+@using Datahub.Portal.Metadata.Components
+@using Datahub.PowerBI.Pages.Admin
+
 
 @using Askmethat.Aspnet.JsonLocalizer.Localizer
 
@@ -37,24 +42,12 @@
 @using Datahub.Core.Layout
 @using Datahub.Achievements
 
-
-@using Datahub.GeoCore.Service
+@using Microsoft.Extensions.Logging;
+@using Microsoft.PowerBI.Api.Models
 
 @using Datahub.Portal.Services.Storage
 @using Datahub.Portal.Services
 @using Datahub.Portal.Components
-@using Datahub.Portal.Components.FileGridComponents
-@using Datahub.Portal.Metadata.Components
-@using Datahub.Portal.Views
-@using Datahub.Portal.Views.Modals
-@using Datahub.Portal.Components.Versions
-@using Datahub.Portal.Components.Forms
-@using Datahub.Portal.Data.Finance;
-@using Datahub.Portal.Data;
-@using Datahub.Portal.Data.LanguageTraining;
-@using Datahub.Portal.Data.WebAnalytics;
-
-@using Datahub.LanguageTraining;
 
 @using Elemental.Components;
