+++ WebPortal/Pages/Project/DataProject/ProjectTools/ProjectDatabase.razor
@@ -1,5 +1,6 @@
 ﻿@*ProjectDatabase.razor*@
 
+@using Datahub.Metadata.Model;
 @using Datahub.Portal.Data.ProjectResource
 @inject IDbContextFactory<DatahubProjectDBContext> _dbFactoryProject
 @inject RequestManagementService _requestManagementService
@@ -259,7 +260,7 @@ else
                             ObjectId=@ResourceID.ToString()
                             Location=@Project.Project_Acronym_CD
                             ProfileName="catalog"
-                            ObjectType=@Metadata.Model.MetadataObjectType.Database
+                            ObjectType=@MetadataObjectType.Database
                             ProjectId=@Project.Project_ID
                             UpdateCatalog
                             ValidateRequired
